engine refinements and tests
- the publisher now run on a separate thread
- main immutable classes
- RingBuffer plumbing between pipelines
- added more tests

+++ src/main/groovy/com/github/fbertola/velociruptor/processing/Plug.groovy
@@ -2,7 +2,7 @@ package com.github.fbertola.velociruptor.processing
 
 trait Plug<T> implements Iterator<T> {
 
-    int expectedSize = 0
+    long expectedSize = 0
 
     abstract void on()
